+++ typescript/prelude.ts
@@ -13,6 +13,7 @@ export const UINT8_MAX = 255
 export const UINT16_MAX = 65535
 export const UINT32_MAX = 4294967295
 
+// Type assertions
 export type PositiveInteger<T extends number> = `${T}` extends '0' | `-${any}` | `${any}.${any}` ? never : T
 export type NonnegativeInteger<T extends number> = `${T}` extends `-${any}` | `${any}.${any}` ? never : T